pyproject.toml
[project]
name = "Inkar-Suki"
version = "0.8.7"
description = "A IM chat bot based on OneBot V11."
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "nonebot2[fastapi]>=2.3.3",
    "nonebot-plugin-alconna>=0.52.2",
    "pillow==10.0.0",
    "nonebot-plugin-apscheduler>=0.5.0",
    "numpy>=2.0.2",
    "playwright>=1.47.0",
    "nonebot-adapter-onebot>=2.4.5",
    "httpx>=0.27.2",
    "pydantic>=2.9.2",
    "simpleeval>=0.9.13",
    "matplotlib>=3.9.2",
    "typing-extensions>=4.12.2",
    "nb-cli>=1.4.2",
    "nonebot-plugin-handle>=0.4.1",
    "beautifulsoup4>=4.12.3",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_alconna", "nonebot_plugin_apscheduler", "nonebot_plugin_handle"]
plugin_dirs = ["src/plugins"]
builtin_plugins = []

[tool.pdm]
